Moved ranges formatting function and added tests
